requirements.txt
langchain>=0.1.16
langchain-google-genai>=1.0.3
langchain-community>=0.0.33
chromadb>=0.4.24
pypdf>=4.2.0
tiktoken>=0.6.0
streamlit>=1.33.0
plotly>=5.20.0
umap-learn>=0.5.5
scikit-learn>=1.4.1.post1
numpy>=1.26.4
pandas>=2.2.1
python-dotenv>=1.0.1